requirements.txt
# 安裝方式 pip install -r requirements.txt
fastapi==0.87.0
# uvicorn[standard] 才有 websocket support
uvicorn[standard]==0.19.0
pydantic==1.10.2
Starlette==0.21.0
line-bot-sdk==2.3.0
python-dotenv==0.21.0
wheel
numpy==1.23.5
# tensorflow==2.11.0
# tensorflow==2.9.0
tensorflow-cpu == 2.11.0
python-multipart==0.0.5
Pillow==9.3.0
# 安裝 cv2 (opencv-python) # 4.3.0.38, 4.4.0.46, 4.5.5.64, 4.6.0.66
# opencv-python==4.4.0.46
opencv-python==4.3.0.38

Jinja2
# aiosqlite==0.17.0
# anyio==3.6.2
# click==8.1.3
# colorama==0.4.6
# h11==0.14.0
# idna==3.4
# iso8601==1.1.0
# pypika-tortoise==0.1.6
# pytz==2022.6
# sniffio==1.2.0
# tortoise-orm==0.19.2
# typing-extensions==4.4.0
# gunicorn==20.1.0
gunicorn

tensorflow_datasets
tensorflow_addons
google-cloud-storage

imutils
firebase_admin